simu_backend: fix compilation warning

+++ host/src/simu_backend.c
@@ -187,7 +187,7 @@ static void align_on_dpu(unsigned int dpu_offset, unsigned rank_id, int pass_id)
 {
     int nb_map = 0;
     int numdpu = dpu_offset + rank_id;
-    if (numdpu > index_get_nb_dpu())
+    if (numdpu > (int)index_get_nb_dpu())
         return;
     int size_neighbour_in_symbols = SIZE_IN_SYMBOLS(delta_neighbour);
     dispatch_request_t *requests = dispatch_get(numdpu, pass_id);
